+++ src/config/updater/from_v0_1.rs
@@ -17,9 +17,7 @@
 
 use toml_edit::{Document, Item, Table};
 
-use crate::config::VERSION;
-
-use super::super::split_type_and_doc;
+use super::{super::split_type_and_doc, common, AskForTicket};
 
 const OBSOLETE_TYPES_DOC: &str = "
 #
@@ -30,11 +28,16 @@ const OLD_SCOPES_DOC: &str = "The list of valid scopes.";
 const NEW_SCOPES_DOC: &str = "The accepted scopes.";
 
 /// Updates the configuration from version 0.1.
-pub fn update(toml_config: &mut Document) {
-    update_version(toml_config);
+pub fn update(toml_config: &mut Document, ask_for_ticket: AskForTicket) {
+    common::update_version(toml_config);
     update_types(toml_config);
     update_scopes(toml_config);
-    update_ticket(toml_config);
+
+    match ask_for_ticket {
+        AskForTicket::Ask { require } => update_ticket(toml_config, require),
+        AskForTicket::DontAsk => remove_ticket(toml_config),
+    }
+
     update_templates(toml_config);
 }
 
@@ -42,11 +45,6 @@ pub fn update(toml_config: &mut Document) {
 // handling. This is because `ConfigUpdater::load` already validates the
 // configuration by parsing it to a `Config`. Any error occuring here is a bug,
 // hence should lead to a panic.
-
-fn update_version(toml_config: &mut Document) {
-    let version = toml_config.get_mut("version").expect("No `version` key");
-    *version = Item::Value(VERSION.into());
-}
 
 fn update_types(toml_config: &mut Document) {
     let doc = toml_config
@@ -99,7 +97,7 @@ fn update_scopes(toml_config: &mut Document) {
     toml_config.insert("scopes", Item::Table(scopes));
 }
 
-fn update_ticket(toml_config: &mut Document) {
+fn update_ticket(toml_config: &mut Document, required: bool) {
     let (key, value) = toml_config
         .get_key_value("ticket_prefixes")
         .expect("No `ticket_prefixes` key");
@@ -112,6 +110,7 @@ fn update_ticket(toml_config: &mut Document) {
         .expect("Improper string in the prefix decorator of the `ticket_prefixes` key");
 
     let mut ticket = Table::new();
+    ticket.insert("required", Item::Value(required.into()));
     ticket.insert("prefixes", value.clone());
     ticket
         .key_decor_mut("prefixes")
@@ -120,6 +119,10 @@ fn update_ticket(toml_config: &mut Document) {
 
     toml_config.remove("ticket_prefixes");
     toml_config.insert("ticket", Item::Table(ticket));
+}
+
+fn remove_ticket(toml_config: &mut Document) {
+    toml_config.remove("ticket_prefixes");
 }
 
 fn update_templates(toml_config: &mut Document) {
@@ -136,8 +139,11 @@ fn update_templates(toml_config: &mut Document) {
             "Improper string in the prefix decorator of the `template` key",
         );
 
+    let template = value.as_str().expect("The `template` key is not a string");
+    let template = common::add_ticket_condition_to_commit_template(template);
+
     let mut templates = Table::new();
-    templates.insert("commit", value.clone());
+    templates.insert("commit", Item::Value(template.into()));
     templates
         .key_decor_mut("commit")
         .expect("No `commit` key")